Fixed sometimes choosing small stacks over large ones
This is an ancient problem in Breakables where the merging mechanism was always combining the same breakable into a single entry in its internal array and disregarding what bag and slot each stack was in. If we retain the position of the largest stack, then Breakables can always pick the proper larger stack to break no matter what order they appear in bags. This won't help the case where a single stack is spread into stacks where no single stack is large enough to break, but the WoW client has built-in bag compression now, so that should be less of a concern.