Added support for combining smaller items into larger ones
Certain types of herbs and ore can be looted that aren't millable/prospectable themselves, but can be used to combine (or unbreak, if you will...) them into an item that can be milled/prospected. It's convenient if the addon offers support for those situations so you can combine the combinables, then break the breakables.