Updated for 8.0 API changes
The equipment manager has been converted to C code, so its functionality (and index basis) changed.
The ilvl squish made the "can this be disenchanted?" function completely wrong and I don't even know if there's an enchanting level restriction anymore, so for now if you're on 8.0 the ilvl check is just bypassed. This is probably going to need refining because it currently, for example, shows legendary items as disenchantable even though the game doesn't allow them to be disenchanted. I'm unsure if it works the same way on Live currently or not.